requirements.txt
antiorm==1.2.1
asgiref==3.8.1
asttokens==2.4.1
audioread==3.0.1
bcrypt==4.1.2
blinker==1.8.2
build==1.0.3
build.py==1.0.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
cluster==1.4.1.post3
cmake==3.28.1
colorama==0.4.6
comm==0.2.0
contourpy==1.2.0
crispy-bootstrap4 @ git+https://github.com/django-crispy-forms/crispy-bootstrap4.git@44eb449f852b78971d55a0b377699b31f44d92ae
customtkinter==5.2.1
cycler==0.12.1
darkdetect==0.8.0
db==0.1.1
db-sqlite3==0.0.1
debugpy==1.8.0
decorator==5.1.1
distlib==0.3.8
dj-database-url==2.3.0
Django==5.1.3
django-cors-headers==4.6.0
django-crispy-forms==2.3
django-filter==24.3
djangorestframework==3.15.2
executing==2.0.1
ffmpeg==1.4
filelock==3.13.1
Flask==3.0.3
fonttools==4.46.0
future==1.0.0
gunicorn==23.0.0
h11==0.14.0
idna==3.10
ipykernel==6.27.1
ipython==8.18.1
itsdangerous==2.2.0
jason==0.1.7
javascript==1!1.1.2
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonpickle==3.3.0
jupyter_client==8.6.0
jupyter_core==5.5.0
kiwisolver==1.4.5
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
MarkupSafe==3.0.2
matplotlib==3.8.2
matplotlib-inline==0.1.6
more-itertools==10.5.0
msgpack==1.1.0
music21==9.3.0
mysqlclient==2.2.6
nest-asyncio==1.5.8
npm==0.1.1
numba==0.60.0
numpy==1.26.2
olefile==0.47
opencv-python==4.8.1.78
optional-django==0.1.0
packaging==23.2
parso==0.8.3
phyton==1.0.0a1
Pillow==10.1.0
pipr==0.0.50
platformdirs==4.0.0
pooch==1.8.2
prompt-toolkit==3.0.41
psutil==5.9.6
psycopg2==2.9.9
psycopg2-binary==2.9.10
pure-eval==0.2.2
py==1.11.0
pycparser==2.22
pydub==0.25.1
pygame==2.5.2
Pygments==2.17.2
PyMySQL==1.1.1
pyparsing==3.1.1
pyproject_hooks==1.0.0
pysqlite3-wheels==0.5.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.2
pywin32==306
pyzmq==25.1.1
requests==2.32.3
running==0.1.3
scikit-learn==1.5.2
scipy==1.14.1
serializers==0.2.4
six==1.16.0
soundfile==0.12.1
soxr==0.5.0.post1
sqlparse==0.4.4
stack-data==0.6.3
threadpoolctl==3.5.0
tornado==6.4
traitlets==5.14.0
typing_extensions==4.12.2
tzdata==2023.4
Unidecode==1.3.8
urllib3==2.2.3
uvicorn==0.32.0
vev==0.5.0
virtualenv==20.25.0
virtualenvwrapper-win==1.2.7
wcwidth==0.2.12
webcolors==24.8.0
Werkzeug==3.0.6
whitenoise==6.8.2
